usb: dwc3-msm: Remove the read of qcom,vbus-present property
Commit 5c0656c667cca76 ("usb: dwc3_otg: Remove logic to manually set BSV")
had removed logic of manually over-riding BSV. But it did not remove the
actual reading of property from the dtsi.
Remove reading of this property completely, as this property is not
supported anymore and handling will be done from userspace for cases
where VBUS indication from PMIC is not present.

Change-Id: Ia761572e8ac096e96685242ffee74bc4d069f9af
Signed-off-by: Azhar Shaikh <azhars@codeaurora.org>