msm: hdcp: proper state sanitization for different versions
The hdcp 2.2 module may or may not be loaded based on source
and sink capabilities. Some of the common functionalities with
hdcp 1.4 may be accessed by different modules depending on which
hdcp version is being used. Check for validity of internal data
structures and variables to guard hdcp functionalities against
illicit calls.

Change-Id: I1383ae77f0b904589bcab571c902b5e3eeef6efa
Signed-off-by: Ajay Singh Parmar <aparmar@codeaurora.org>