platform: pon-reset-off: Store device shut down in PMIC register
Use the control register in PMIC power-on peripheral store device
in the shut down state. When the power off button was pushed and from
the UI were clicked on "shut down", in the case of this particular flag:
ON means that the device will stay in shut down state instead of shut down
and boot device again.

Change-Id: I2198a37427a80c51fffcec47d9476892cdc662c5
Signed-off-by: Tanya Finkel <tfinkel@codeaurora.org>