soc: qcom: glink_smd_xprt: Defer channel close ACK operation
Depending on the timing of two competing transport registration with
G-Link core, a channel gets opened on one transport and then immediately
migrated to another transport. This channel migration happens in the
context of glink_open() operation itself and causes a re-lock attempt
on an already locked mutex.

Defer the channel close ACK operation so that channel migration does not
hit the deadlock.

CRs-Fixed: 1004150
Change-Id: I188846b95369b674830bc01ddeca764ad6d4d391
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>